perf(tests): cut per-file import graphs in apps/sim and shard the suite across CI runners
