+++ apps/sim/lib/copilot/tools/server/blocks/get-blocks-metadata-projection.test.ts
@@ -30,6 +30,13 @@ vi.mock('@/lib/integrations/availability.server', () => ({
 
 import { getBlocksMetadataServerTool } from '@/lib/copilot/tools/server/blocks/get-blocks-metadata-tool'
 
+/**
+ * The projection under test reads real tool params and outputs, which the global
+ * `@/tools/metadata` and `@/tools/metadata-outputs` mocks in vitest.setup.ts empty.
+ */
+vi.unmock('@/tools/metadata')
+vi.unmock('@/tools/metadata-outputs')
+
 interface AgentBlockMetadata {
   blockType: string
   name: string
