+++ apps/sim/tools/windchill/registry.test.ts
@@ -2,17 +2,14 @@
  * @vitest-environment node
  */
 
-import { describe, expect, it, vi } from 'vitest'
-
-vi.unmock('@/tools/registry')
-
-import { tools } from '@/tools/registry'
+import { describe, expect, it } from 'vitest'
+import { hasToolId } from '@/tools/tool-ids'
 import { WINDCHILL_OPERATIONS } from '@/tools/windchill/types'
 
 describe('Windchill registry', () => {
   it('registers every operation in the global tool registry', () => {
     for (const operation of WINDCHILL_OPERATIONS) {
-      expect(tools[operation]?.id).toBe(operation)
+      expect(hasToolId(operation), operation).toBe(true)
     }
   })
 })
