+++ package.json
@@ -14,19 +14,8 @@
     "dev:sockets": "cd apps/realtime && bun run dev",
     "dev:full": "bunx concurrently -n \"App,Realtime\" -c \"cyan,magenta\" \"cd apps/sim && bun run dev\" \"cd apps/realtime && bun run dev\"",
     "dev:full:capped": "bunx concurrently -n \"App,Realtime\" -c \"cyan,magenta\" \"cd apps/sim && bun run dev:capped\" \"cd apps/realtime && bun run dev\"",
-    "test": "bun run test:setup && bun run test:npm-package-versions && bun run test:audit-candidates && bun run test:icon-path-precision && bun run test:tool-registry-boundary && bun run test:tool-request-boundary && bun run test:actorless-executor-operations && bun run test:permission-group-enforcement && bun run test:capability-subject && bun run test:application-graph && bun run test:migrations-safety && bun run test:generators && turbo run test",
+    "test": "bun run test:setup && bun run test:scripts && turbo run test",
     "test:setup": "bun run --cwd packages/sim-setup test",
-    "test:npm-package-versions": "bunx vitest run scripts/bump-npm-package-versions.test.ts",
-    "test:audit-candidates": "bunx vitest run scripts/check-db-audit-candidates.test.ts scripts/check-egress-boundary.test.ts",
-    "test:icon-path-precision": "bunx vitest run scripts/check-icon-path-precision.test.ts",
-    "test:tool-registry-boundary": "bunx vitest run scripts/check-tool-registry-boundary.test.ts",
-    "test:tool-request-boundary": "bunx vitest run scripts/check-tool-request-boundary.test.ts",
-    "test:actorless-executor-operations": "bunx vitest run scripts/check-actorless-executor-operations.test.ts",
-    "test:permission-group-enforcement": "bunx vitest run scripts/check-permission-group-enforcement.test.ts",
-    "test:capability-subject": "bunx vitest run scripts/check-capability-subject.test.ts",
-    "test:application-graph": "bunx vitest run scripts/check-application-graph.test.ts",
-    "test:migrations-safety": "bunx vitest run scripts/check-migrations-safety.test.ts",
-    "test:generators": "bunx vitest run scripts/generate-v2-cli-api.test.ts scripts/generate-cli-docs.test.ts scripts/generate-docs.test.ts scripts/generate-block-successors.test.ts",
     "format": "turbo run format",
     "format:check": "turbo run format:check",
     "lint": "turbo run lint",
@@ -119,7 +108,8 @@
     "agent-stream-docs:check": "bun run scripts/sync-agent-stream-docs.ts --check",
     "prepare": "bun husky",
     "type-check": "turbo run type-check",
-    "release": "bun run scripts/create-single-release.ts"
+    "release": "bun run scripts/create-single-release.ts",
+    "test:scripts": "vitest run --config vitest.scripts.config.ts"
   },
   "overrides": {
     "react": "19.2.4",
