+++ packages/testing/src/factories/tool-responses.factory.ts
@@ -4,8 +4,6 @@
  * This file contains mock data samples to be used in tool unit tests.
  */
 
-import { randomFloat } from '@sim/utils/random'
-
 /**
  * HTTP Request mock responses for different scenarios.
  */
@@ -147,9 +145,7 @@ export const mockSheetsResponses = {
  */
 export const mockPineconeResponses = {
   embedding: {
-    embedding: Array(1536)
-      .fill(0)
-      .map(() => randomFloat() * 2 - 1),
+    embedding: Array.from({ length: 1536 }, (_, i) => ((i * 7919) % 2000) / 1000 - 1),
     metadata: { text: 'Sample text for embedding', id: 'embed-123' },
   },
   searchResults: {
